dev articles form, model

+++ accounts/views.py
@@ -26,7 +26,8 @@ def login(request):
 
 @login_required
 def logout(request):
-    auth_logout(request)
+    if request.user.is_authenticated:
+        auth_logout(request)
     return redirect('articles:index')
 
 
@@ -52,8 +53,9 @@ def signup(request):
 
 
 def delete(request):
-    request.user.delete()
-    auth_logout(request)
+    if request.user.is_authenticated:
+        request.user.delete()
+        auth_logout(request)
     return redirect('articles:index')
